Creation of classes for specific functionality
