+++ .gitignore
@@ -1,3 +1,2 @@
 /node_modules
-app.js
-app.js.map
+/dist
