+++ index.html
@@ -2,8 +2,9 @@
 <html>
 <head>
 	<title>JS Onboard Project</title>
+	<link rel="stylesheet" href="stylesheet.css">
 	<script type="text/javascript" charset="utf-8" src="third_party/jquery-2.0.3.min.js"></script>
-	<script type="text/javascript" charset="utf-8" src="app.js"></script>
+	<script type="module" charset="utf-8" src="dist/app.js"></script>
 </head>
 
 <body>
