+++ package.json
@@ -2,10 +2,9 @@
   "name": "onboard-javascript",
   "version": "1.0.0",
   "description": "This is a JavaScript project for all new developers to complete before venturing into our web frontend codebase.",
-  "main": "index.js",
+  "main": "dist/index.js",
   "scripts": {
-    "build": "tsc --build",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "build": "tsc --build"
   },
   "repository": {
     "type": "git",
