+++ stylesheet.css
@@ -1,0 +1,9 @@
+* {box-sizing: border-box; 
+    margin: 0;
+    padding: 0;
+}
+
+body {width: 100vw;
+    height: 100vh;
+    overflow: hidden;
+}
